Refactor : `domain\promotion` 패키지 생성
