+++ src/main/java/christmas/Application.java
@@ -1,11 +1,11 @@
 package christmas;
 
 import camp.nextstep.edu.missionutils.Console;
-import christmas.controller.Planner;
+import christmas.controller.PlannerController;
 
 public class Application {
     public static void main(String[] args) {
-        Planner eventPlaner = new Planner();
+        PlannerController eventPlaner = new PlannerController();
         eventPlaner.run();
         Console.close();
     }
