--- src/main/java/christmas/exception/ChristmasException.java
@@ -1,7 +1,0 @@
-package christmas.exception;
-
-public class ChristmasException extends IllegalArgumentException {
-    ChristmasException(String message) {
-        super(message);
-    }
-}
